# Optimize resizeBilinearBackprop on CPU