#### Description
<!--
Please describe the pull request here.
Also, if this is an issue/bug fix, please add the issue link for reference here.
-->
Optimize the implementation of `resizeBilinearBackprop` in `backend_cpu.ts`

Benchmarking code, to run you need to expose `resizeBilinearBackprop` as a top level op:
```
tf.setBackend("cpu");
const ns = [64, 128, 256, 512];
const RUNS = 100;

ns.forEach(n => {
    const t = tf.randomUniform([8, n/2, n/2, 3]);
    const dy = tf.randomUniform([8, n, n, 3]);
    // Warmup.
    tf.image.resizeBilinearBackprop(dy, t, false);

    let res = null;
    const start = performance.now();
    for (let i = 0; i < RUNS; i++) {
        res = tf.image.resizeBilinearBackprop(dy, t, false);
    }
    res.dataSync();
    const elapsed = (performance.now() - start) / RUNS;
    console.log(`N: ${n}: ${elapsed.toFixed(2)} ms`);
});
```
Results before optimization:
```
N: 64: 76.09 ms
N: 128: 290.47 ms
N: 256: 1207.47 ms
N: 512: 4663.63 ms
```

Results after optimization:
```
N: 64: 1.16 ms
N: 128: 4.26 ms
N: 256: 17.38 ms
N: 512: 67.65 ms
```


---
<!-- Please do not delete this section -->
##### For repository owners only:

Please remember to apply all applicable tags to your pull request.
Tags: FEATURE, BREAKING, BUG, PERF, DEV, DOC, SECURITY

For more info see: https://github.com/tensorflow/tfjs/blob/master/DEVELOPMENT.md

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-core/1299)
<!-- Reviewable:end -->